# specs/openiap-kit/generated/openapi/commerce-protocol.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenIAP Commerce Protocol",
    "version": "1.0",
    "description": "REST binding of the OpenIAP Commerce Protocol operation surface. Generated from commerce-protocol.graphql — do not edit. Error statuses follow the manifest's errorStatus table; every error body is a ProtocolErrorResponse.",
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
    "/commerce/v1/capabilities": {
      "get": {
        "operationId": "providerCapabilities",
        "summary": "What this provider implements: protocol version, profiles, bindings, and per-store capability",
        "description": "What this provider implements: protocol version, profiles, bindings, and per-store capability. Public and free of commerce data, so a consumer or a conformance runner reads it without credentials and without any central registry.",
        "tags": [
          "core"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderCapabilities"
                }
              }
            }
          },
          "429": {
            "description": "`RATE_LIMITED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`INTERNAL_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/commerce/v1/subscriptions/status": {
      "get": {
        "operationId": "subscriptionStatus",
        "summary": "A developer backend reads one authenticated user's subscription standing",
        "description": "A developer backend reads one authenticated user's subscription standing. Server role only: a shipped app must not be able to walk arbitrary user identities, which is why the verification role is refused here.",
        "tags": [
          "entitlements"
        ],
        "security": [
          {
            "serverCredential": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionStatusResult"
                }
              }
            }
          },
          "400": {
            "description": "`INVALID_REQUEST`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`RATE_LIMITED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`INTERNAL_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "description": "The opaque app-scoped user identity the caller's backend selected, in the identity space shared by provider and caller.",
              "$ref": "#/components/schemas/Identifier"
            }
          }
        ]
      }
    },
    "/commerce/v1/entitlements": {
      "get": {
        "operationId": "entitlements",
        "summary": "A developer backend reads one authenticated user's access decision",
        "description": "A developer backend reads one authenticated user's access decision. Server role only, and fail-close: anything the provider cannot classify as currently entitled is absent from the answer.",
        "tags": [
          "entitlements"
        ],
        "security": [
          {
            "serverCredential": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitlementsResult"
                }
              }
            }
          },
          "400": {
            "description": "`INVALID_REQUEST`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`RATE_LIMITED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`INTERNAL_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "description": "The opaque app-scoped user identity the caller's backend selected, in the identity space shared by provider and caller.",
              "$ref": "#/components/schemas/Identifier"
            }
          }
        ]
      }
    },
    "/commerce/v1/purchases/verify": {
      "post": {
        "operationId": "verifyPurchase",
        "summary": "Verifies store evidence and returns the verdict",
        "description": "Verifies store evidence and returns the verdict. The verification role suffices: this operation binds no user, reads no account, and repeating it with the same evidence is safe. VERIFICATION_FAILED means the provider could not obtain a verdict — it never stands in for the store rejecting the evidence, which is a successful result with `isValid: false`.",
        "tags": [
          "verification"
        ],
        "security": [
          {
            "verificationCredential": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyPurchaseResult"
                }
              }
            }
          },
          "400": {
            "description": "`INVALID_REQUEST`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "`UNSUPPORTED_STORE`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`RATE_LIMITED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`INTERNAL_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "`VERIFICATION_FAILED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPurchaseInput"
              }
            }
          }
        }
      }
    },
    "/commerce/v1/purchases/bind": {
      "post": {
        "operationId": "bindPurchase",
        "summary": "Binds a verified purchase to the adopter's opaque user identity",
        "description": "Binds a verified purchase to the adopter's opaque user identity. Server role only; retrying is idempotent.",
        "tags": [
          "accountLifecycle"
        ],
        "security": [
          {
            "serverCredential": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BindPurchaseResult"
                }
              }
            }
          },
          "400": {
            "description": "`INVALID_REQUEST`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "`UNSUPPORTED_STORE`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`RATE_LIMITED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`INTERNAL_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindPurchaseInput"
              }
            }
          }
        }
      }
    },
    "/commerce/v1/users/erase": {
      "post": {
        "operationId": "eraseUser",
        "summary": "Removes a user identity from the provider's subscription records and protocol event identity",
        "description": "Removes a user identity from the provider's subscription records and protocol event identity. Server role only; re-requesting the same user is idempotent and reports the current job.",
        "tags": [
          "accountLifecycle"
        ],
        "security": [
          {
            "serverCredential": []
          }
        ],
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EraseUserResult"
                }
              }
            }
          },
          "400": {
            "description": "`INVALID_REQUEST`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`RATE_LIMITED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`INTERNAL_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EraseUserInput"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProviderCapabilities": {
        "title": "OpenIAP Commerce Protocol implementation descriptor",
        "description": "What an implementation supports: the specification version it speaks, the event types it can emit, and what it can actually observe per store. This is the document a consumer, an operator, or a tool reads to determine compatibility without guessing and without reading prose. It contains no commerce data.",
        "type": "object",
        "required": [
          "specVersion",
          "eventTypes",
          "stores"
        ],
        "additionalProperties": true,
        "properties": {
          "specVersion": {
            "description": "OpenIAP Commerce Protocol version this declaration was written against, as MAJOR.MINOR.",
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)\\.(0|[1-9][0-9]*)$"
          },
          "implementation": {
            "description": "Free-form name and version of the backend making this declaration. Present so an operator can tell two backends apart; it carries no normative meaning.",
            "type": "object",
            "additionalProperties": true,
            "properties": {
              "name": {
                "description": "Name of the backend making this declaration, for an operator reading a descriptor without knowing where it came from.",
                "type": "string",
                "minLength": 1
              },
              "version": {
                "description": "Version of that backend, when it publishes one. Not a specification version — that is `specVersion`.",
                "type": "string",
                "minLength": 1
              }
            }
          },
          "eventTypes": {
            "description": "Event types this implementation can emit. A consumer uses it to know which lifecycle signals to expect; absence of a type here means this implementation never produces it, which is different from a type that simply has not occurred yet.",
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[a-z]+\\.[a-z_]+$"
            },
            "minItems": 1,
            "uniqueItems": true
          },
          "stores": {
            "description": "Keyed by store, using the same open value space as the event envelope. An implementation declares only the stores it actually integrates.",
            "type": "object",
            "minProperties": 1,
            "propertyNames": {
              "pattern": "^[a-z][a-z0-9_]*$"
            },
            "additionalProperties": {
              "$ref": "#/components/schemas/CapabilitiesStoreCapabilities"
            }
          },
          "profiles": {
            "description": "Operation profiles this implementation serves, keyed by profile name with the profile version as the value. This version names `verification`, `entitlements`, `events`, and `accountLifecycle`; the key space is open, so a consumer MUST ignore a profile it does not recognise. An implementation MUST declare only profiles it implements and passes conformance for, and MUST NOT declare a profile it partially implements. Absent on a descriptor from an events-only emitter that predates the operation surface.",
            "type": "object",
            "minProperties": 1,
            "propertyNames": {
              "pattern": "^[a-z][a-zA-Z0-9]*$"
            },
            "additionalProperties": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]*)\\.(0|[1-9][0-9]*)$"
            }
          },
          "bindings": {
            "description": "Transport bindings this implementation serves for the declared profiles, keyed by binding name with the binding version as the value. This version names `rest` and `graphql`; the key space is open. Declaring a binding means every declared profile operation is reachable over it. An implementation MAY serve one binding only. Absent on a descriptor from an events-only emitter.",
            "type": "object",
            "minProperties": 1,
            "propertyNames": {
              "pattern": "^[a-z][a-zA-Z0-9]*$"
            },
            "additionalProperties": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]*)\\.(0|[1-9][0-9]*)$"
            }
          }
        }
      },
...
      "SubscriptionStatusResult": {
        "description": "The user's current subscription standing. `active` answers the gate for the user as a whole; `subscription` is the most relevant record — the current entitling subscription when one exists, otherwise the provider's most recent record as context. It is omitted when the provider has no record for this user. A provider that cannot enumerate the user's records completely MUST fail the operation instead of answering from a partial read.",
        "type": "object",
        "required": [
          "active"
        ],
        "additionalProperties": false,
        "properties": {
          "active": {
            "description": "Whether the user is entitled to anything right now. The gate for simple access checks.",
            "type": "boolean"
          },
          "subscription": {
            "$ref": "#/components/schemas/SubscriptionStatusSnapshot"
          }
        }
      },
...
      "EntitlementsResult": {
        "description": "Every product the user may access right now, with the entitling subscription records. Unknown, expired, and ambiguous records contribute nothing: a product appears only when its gate is open at the provider's read time. A provider that cannot enumerate the user's records completely MUST fail the operation instead of answering from a partial read.",
        "type": "object",
        "required": [
          "userId",
          "productIds",
          "subscriptions"
        ],
        "additionalProperties": false,
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/Identifier"
          },
          "productIds": {
            "description": "Every product the user may access right now, deduplicated.",
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "uniqueItems": true
          },
          "subscriptions": {
            "description": "The subscription records whose open gates produced `productIds`.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubscriptionStatusSnapshot"
            }
          }
        }
      },
...
      "BindPurchaseResult": {
        "description": "Whether a binding now exists between this user and the evidenced purchase. `bound: false` covers every non-binding outcome — unknown evidence, evidence bound to a different user, or a store the provider cannot bind — without distinguishing them, so the operation cannot be used to probe whether someone else's purchase exists. Re-binding the same user to the same evidence is idempotent and reports `bound: true`.",
        "type": "object",
        "required": [
          "bound"
        ],
        "additionalProperties": true,
        "properties": {
          "bound": {
            "description": "True when a binding between this user and the evidenced purchase now exists.",
            "type": "boolean"
          }
        }
      },
      "EraseUserInput": {
        "description": "Selects the user identity to erase from the provider's records.",
        "type": "object",
        "required": [
          "userId"
        ],
        "additionalProperties": true,
        "properties": {
          "userId": {
            "description": "The opaque app-scoped user identity to remove from the provider's subscription records and protocol event identity.",
            "$ref": "#/components/schemas/Identifier"
          }
        }
      },
...
      "ProtocolErrorResponse": {
        "description": "The REST error envelope. Every non-success REST response carries exactly this shape — CLOSED, so an error body cannot smuggle members past the tokenless rules (a failure response is the easiest place to hide a leak, because callers rarely inspect it).",
        "type": "object",
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ProtocolError"
          }
        }
      },
...
      "Identifier": {
        "description": "An opaque, implementation-assigned identifier. Consumers MUST treat it as an opaque string and MUST NOT parse structure out of it.",
        "type": "string",
        "minLength": 1,
        "maxLength": 256
      },
...
    },
    "securitySchemes": {
      "verificationCredential": {
        "type": "http",
        "scheme": "bearer",
        "description": "Verification-role credential: limited to purchase-evidence verification, no account access."
      },
      "serverCredential": {
        "type": "http",
        "scheme": "bearer",
        "description": "Server-role credential for an authenticated developer backend. Never ship it in an app."
      }
    }
  },
...
}